requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.12.2
pytest==6.2.5
pytest-django==4.5.2
